getting rid of the nav bar
