+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/announcements/AnnouncementsListActivity.java
@@ -5,13 +5,13 @@ import android.os.Bundle;
 import blueprint.com.sage.R;
 import blueprint.com.sage.main.fragments.AnnouncementsListFragment;
 import blueprint.com.sage.network.Requests;
-import blueprint.com.sage.shared.activities.NavigationAbstractActivity;
+import blueprint.com.sage.shared.activities.BackAbstractActivity;
 import blueprint.com.sage.utility.view.FragUtils;
 
 /**
  * Created by kelseylam on 10/24/15.
  */
-public class AnnouncementsListActivity extends NavigationAbstractActivity {
+public class AnnouncementsListActivity extends BackAbstractActivity {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
