+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/BrowseActivity.java
@@ -10,7 +10,7 @@ import blueprint.com.sage.browse.fragments.BrowseTabFragment;
 import blueprint.com.sage.models.School;
 import blueprint.com.sage.models.User;
 import blueprint.com.sage.network.Requests;
-import blueprint.com.sage.shared.activities.NavigationAbstractActivity;
+import blueprint.com.sage.shared.activities.BackAbstractActivity;
 import blueprint.com.sage.shared.interfaces.SchoolsInterface;
 import blueprint.com.sage.shared.interfaces.UsersInterface;
 import blueprint.com.sage.utility.view.FragUtils;
@@ -18,7 +18,7 @@ import blueprint.com.sage.utility.view.FragUtils;
 /**
  * Created by charlesx on 11/4/15.
  */
-public class BrowseActivity extends NavigationAbstractActivity implements SchoolsInterface, UsersInterface {
+public class BrowseActivity extends BackAbstractActivity implements SchoolsInterface, UsersInterface {
 
     private List<School> mSchools;
     private List<User> mUsers;
