+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/fragments/BrowseTabFragment.java
@@ -4,7 +4,6 @@ import android.os.Bundle;
 
 import blueprint.com.sage.shared.adapters.SimplePagerAdapter;
 import blueprint.com.sage.shared.fragments.TabsFragment;
-import blueprint.com.sage.shared.interfaces.NavigationInterface;
 
 /**
  * Created by charlesx on 11/17/15.
@@ -13,12 +12,9 @@ public class BrowseTabFragment extends TabsFragment {
 
     private SimplePagerAdapter mAdapter;
 
-    private NavigationInterface mNavigationInterface;
-
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        mNavigationInterface = (NavigationInterface) getActivity();
     }
 
     public static BrowseTabFragment newInstance() { return new BrowseTabFragment(); }
@@ -32,7 +28,6 @@ public class BrowseTabFragment extends TabsFragment {
         mViewPager.setAdapter(mAdapter);
         mTabLayout.setupWithViewPager(mViewPager);
 
-        mNavigationInterface.toggleDrawerUse(true);
         getActivity().setTitle("Browse");
     }
 }
