+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/fragments/CreateAdminFragment.java
@@ -22,7 +22,7 @@ public class CreateAdminFragment extends UserFormAbstractFragment {
         mPassword.setHint(getString(R.string.sign_up_password));
         mConfirmPassword.setHint(getString(R.string.sign_up_password_confirm));
 
-        mNavigationInterface.toggleDrawerUse(false);
+        mBackInterface.toggleDrawerUse(false);
         getActivity().setTitle("Create User");
     }
