+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/fragments/EditSchoolFragment.java
@@ -23,7 +23,7 @@ public class EditSchoolFragment extends SchoolFormAbstractFragment {
         mSchoolAddress.setText(mSchool.getAddress());
         setUserSpinner();
 
-        mNavigationInterface.toggleDrawerUse(false);
+        mBackInterface.toggleDrawerUse(false);
         getActivity().setTitle("Edit School");
     }
