+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/browse/fragments/EditUserFragment.java
@@ -87,7 +87,7 @@ public class EditUserFragment extends UserFormAbstractFragment {
         try {
             NetworkUtils.setUser(getActivity(), event.getUser());
             mBaseInterface.setUser(event.getUser());
-            mNavigationInterface.initializeUser();
+            mBackInterface.initializeUser();
         } catch(Exception e) {
             Log.e(getClass().toString(), e.toString());
         }
