+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/requests/fragments/VerifyUsersListFragment.java
@@ -33,7 +33,6 @@ public class VerifyUsersListFragment extends Fragment implements OnRefreshListen
     private VerifyUserListAdapter mUserAdapter;
     private UsersInterface mUsersInterface;
 
-
     public static VerifyUsersListFragment newInstance() { return new VerifyUsersListFragment(); }
 
     @Override
