+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/shared/activities/BackAbstractActivity.java
@@ -1,0 +1,44 @@
+package blueprint.com.sage.shared.activities;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.support.v7.widget.Toolbar;
+
+import blueprint.com.sage.R;
+import blueprint.com.sage.utility.view.ViewUtils;
+import butterknife.Bind;
+import butterknife.ButterKnife;
+
+/**
+ * Created by charlesx on 11/4/15.
+ * Activity that basically adds a nav bar to your activity;
+ */
+public class BackAbstractActivity extends AbstractActivity {
+
+    @Bind(R.id.toolbar) Toolbar mToolbar;
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_back);
+        ButterKnife.bind(this);
+
+        initializeViews();
+    }
+
+    private void initializeViews() {
+        setSupportActionBar(mToolbar);
+
+        if (getSupportActionBar() != null) {
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+            getSupportActionBar().setDisplayShowHomeEnabled(true);
+        }
+    }
+
+    @Override
+    public void onBackPressed() {
+        Fragment fragment = getSupportFragmentManager().findFragmentById(R.id.container);
+        ViewUtils.hideKeyboard(fragment);
+        super.onBackPressed();
+    }
+}
