--- sage-android/Sage/app/src/main/java/blueprint/com/sage/shared/interfaces/NavigationInterface.java
@@ -1,9 +1,0 @@
-package blueprint.com.sage.shared.interfaces;
-
-/**
- * Created by charlesx on 11/23/15.
- */
-public interface NavigationInterface {
-    void toggleDrawerUse(boolean showDrawers);
-    void initializeUser();
-}
