+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/utility/view/FragUtils.java
@@ -1,5 +1,7 @@
 package blueprint.com.sage.utility.view;
 
+import android.app.Activity;
+import android.content.Intent;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentTransaction;
@@ -26,5 +28,21 @@ public class FragUtils {
 
     public static void popBackStack(FragmentActivity activity) {
         activity.getSupportFragmentManager().popBackStack();
+    }
+
+    public static void startActivity(Activity activity, Class<?> cls) {
+        Intent intent = new Intent(activity, cls);
+        activity.startActivity(intent);
+        activity.overridePendingTransition(0, 0);
+    }
+
+    public static void startActivityBackStack(Activity activity, Class<?> cls) {
+        Intent intent = new Intent(activity, cls);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK |
+                Intent.FLAG_ACTIVITY_CLEAR_TASK |
+                Intent.FLAG_ACTIVITY_NO_ANIMATION);
+
+        activity.startActivity(intent);
+        activity.overridePendingTransition(0, 0);
     }
 }
